safer API.
