+++ crates/rules/src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(incomplete_features)]
 #![feature(inherent_associated_types)]
+#![feature(generic_const_exprs)]
 
 mod matrices;
 mod move_to_sargon;
