+++ crates/rules/src/roles/builder/mod.rs
@@ -1,3 +1,6 @@
+mod confirmation_roles_builder_unit_tests;
+mod primary_roles_builder_unit_tests;
+mod recovery_roles_builder_unit_tests;
 mod roles_builder;
 mod roles_builder_unit_tests;
