+++ crates/rules/src/roles/general_role_with_hierarchical_deterministic_factor_instances.rs
@@ -1,2 +1,1 @@
-
 pub struct GeneralRoleWithHierarchicalDeterministicFactorInstances {}
